Move 1 func to `quit_game.py`: quit_game
