+++ scripts/housekeeping/quit_game.py
@@ -1,0 +1,23 @@
+from sys import exit as sys_exit
+
+import pygame
+
+from scripts.game_structure import game
+from scripts.game_structure.game.settings import game_settings_save
+
+
+def quit_game(savesettings=False, clearevents=False):
+    """
+    Quits the game, avoids a bunch of repeated lines
+    """
+    if savesettings:
+        game_settings_save(None)
+    if clearevents:
+        game.cur_events_list.clear()
+    game.rpc.close_rpc.set()
+    game.rpc.update_rpc.set()
+    pygame.display.quit()
+    pygame.quit()
+    if game.rpc.is_alive():
+        game.rpc.join(1)
+    sys_exit()
